fix: Handle nested gRPC calls better
If in the course of handling one gRPC call another is made, the hubs
transaction name would be clobbered by the most recent gRPC calls
FullMethod. This appends the root transaction name to the gRPC metadata
so that we can send the root transaction name as the original call being
handled that resulted in the later spans.